setup namespaces
