+++ Runtime/HDRP/OitRendererHDRP.cs
@@ -2,17 +2,9 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-public class OitRendererHDRP : MonoBehaviour
+namespace OrderIndependentTransparency.HDRP
 {
-    // Start is called before the first frame update
-    void Start()
+    public class OitRendererHDRP
     {
-        
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-        
     }
 }
