+++ Runtime/Scripts/OitInterface.cs
@@ -1,13 +1,16 @@
 using UnityEngine.Rendering;
 
-public enum OitMode
+namespace OrderIndependentTransparency
 {
-    LinkedList,
-}
+    public enum OitMode
+    {
+        LinkedList,
+    }
 
-public interface IOrderIndependentTransparency
-{
-    void PreRender(CommandBuffer command);
-    void Render(CommandBuffer command, RenderTargetIdentifier src, RenderTargetIdentifier dest);
-    void Release();
+    public interface IOrderIndependentTransparency
+    {
+        void PreRender(CommandBuffer command);
+        void Render(CommandBuffer command, RenderTargetIdentifier src, RenderTargetIdentifier dest);
+        void Release();
+    }
 }
